+++ CHANGELOG.md
@@ -89,6 +89,8 @@ for versions < 1.
 
 * `--console-log-command` default to `on`.
 
+* `Timer` log now changed to more general `Status`.
+
 ### Added
 * Option `--edges` (toml: legend `edges`) for declaring dependencies
   between commands e.g. `shrun cmd1 cmd2 --edges "1 & 2"` means `cmd2` will